list work is separated
