+++ src/routes/index.js
@@ -1,12 +1,12 @@
-import styles from './Routes.module.scss'
-import TodoList from './TodoList'
+import styles from './Routes.module.scss';
+import TodoList from './TodoList';
 
 function App() {
   return (
     <div className={styles.app}>
       <TodoList />
     </div>
-  )
+  );
 }
 
-export default App
+export default App;
